requirements.txt
beautifulsoup4==4.11.1
Flask==2.2.2
fuzzywuzzy==0.18.0
loguru==0.7.0
mysql_connector_repackaged==0.3.1
numpy==1.23.5
pandas==1.5.3
PyYAML==6.0
scikit_learn==1.0.2
